+++ src/Select.tsx
@@ -11,6 +11,14 @@ import { Button } from './Button'
 
 // TODO Attention: scrolling does not work because apparently overflow is set to "hidden" on the body
 
+interface ClassName {
+  root?: string
+  trigger?: string
+  viewport?: string
+  item?: string
+  text?: string
+}
+
 export interface Item {
   value: string
   disabled?: boolean // disabled React select
@@ -19,25 +27,53 @@ export interface Item {
 }
 
 export interface SelectProps {
+  name: string
   items: Item[]
   onChange: (newValue: string) => void
   value?: string
   disabled?: boolean
   size?: 'md' | 'sm'
-  className?: {
-    root?: string
-    trigger?: string
-    viewport?: string
-    item?: string
-    text?: string
-  }
+  className?: ClassName
+  placeholder?: string
+}
+
+interface SelectItemProps {
+  className?: ClassName
+  label: string
+  size?: string
+  value: string
 }
 
 const defaultProps = {
   disabled: false,
   size: 'md',
   className: {},
+  placeholder: undefined,
 }
+
+const SelectItem = React.forwardRef(
+  ({ className, label, size, ...props }: SelectItemProps, forwardedRef) => {
+    return (
+      <RadixSelect.Item
+        className={twMerge(
+          'relative flex items-center px-8 py-2 rounded-md text-gray-700 dark:text-gray-300 font-medium focus:bg-gray-100 dark:focus:bg-gray-900',
+          'rdx-disabled:opacity-50 focus:outline-none select-none',
+          size === 'sm' && 'px-7 text-sm',
+          className?.item
+        )}
+        {...props}
+        ref={forwardedRef as React.ForwardedRef<HTMLDivElement>}
+      >
+        <RadixSelect.ItemText className={twMerge(className?.text)}>
+          {label}
+        </RadixSelect.ItemText>
+        <RadixSelect.ItemIndicator className="absolute inline-flex items-center left-2">
+          <FontAwesomeIcon icon={faCheck} size={size === 'sm' ? 'sm' : '1x'} />
+        </RadixSelect.ItemIndicator>
+      </RadixSelect.Item>
+    )
+  }
+)
 
 export function Select({
   items,
@@ -46,6 +82,8 @@ export function Select({
   disabled,
   size,
   className,
+  name,
+  placeholder,
 }: SelectProps) {
   const [open, setOpen] = useState(false)
 
@@ -56,17 +94,17 @@ export function Select({
   return (
     <div className={twMerge('relative flex', className?.root)}>
       <RadixSelect.Root
-        defaultValue={items[0].value}
+        name={name}
         onValueChange={onChange}
         onOpenChange={(open) => setOpen(open)}
         value={value}
       >
-        <RadixSelect.Trigger asChild>
+        <RadixSelect.Trigger asChild id={name}>
           <Button
             disabled={disabled}
             className={twMerge(size === 'sm' && '!text-sm', className?.trigger)}
           >
-            {shortValue || <RadixSelect.Value />}
+            {shortValue || <RadixSelect.Value placeholder={placeholder} />}
 
             <RadixSelect.Icon
               className={twMerge('ml-2', size === 'sm' && 'ml-0.5')}
@@ -91,31 +129,9 @@ export function Select({
               className?.viewport
             )}
           >
-            <RadixSelect.Group>
-              {items.map((item, ix) => (
-                <RadixSelect.Item
-                  disabled={item.disabled}
-                  key={ix}
-                  value={item.value}
-                  className={twMerge(
-                    'relative flex items-center px-8 py-2 rounded-md text-gray-700 dark:text-gray-300 font-medium focus:bg-gray-100 dark:focus:bg-gray-900',
-                    'rdx-disabled:opacity-50 focus:outline-none select-none',
-                    size === 'sm' && 'px-7 text-sm',
-                    className?.item
-                  )}
-                >
-                  <RadixSelect.ItemText className={twMerge(className?.text)}>
-                    {item.label}
-                  </RadixSelect.ItemText>
-                  <RadixSelect.ItemIndicator className="absolute inline-flex items-center left-2">
-                    <FontAwesomeIcon
-                      icon={faCheck}
-                      size={size === 'sm' ? 'sm' : '1x'}
-                    />
-                  </RadixSelect.ItemIndicator>
-                </RadixSelect.Item>
-              ))}
-            </RadixSelect.Group>
+            {items.map((item, ix) => (
+              <SelectItem key={ix} {...item} className={className} />
+            ))}
           </RadixSelect.Viewport>
           <RadixSelect.ScrollDownButton className="flex items-center justify-center text-gray-700 dark:text-gray-300">
             <FontAwesomeIcon